separated url-search params and url classes and added more tests
