+++ packages/svelte/src/reactivity/index-client.js
@@ -1,7 +1,8 @@
 export { SvelteDate } from './date.js';
 export { SvelteSet } from './set.js';
 export { SvelteMap } from './map.js';
-export { SvelteURL, SvelteURLSearchParams } from './url.js';
+export { SvelteURL } from './url.js';
+export { SvelteURLSearchParams } from './url-search-params.js';
 
 /** @deprecated Use `SvelteDate` instead */
 export function Date() {
