+++ vitest.config.js
@@ -30,9 +30,9 @@ export default defineConfig({
 		dir: '.',
 		reporters: ['dot'],
 		include: [
-			'packages/svelte/**/*.test.ts',
-			'packages/svelte/tests/*/test.ts',
-			'packages/svelte/tests/runtime-browser/test-ssr.ts'
+			'packages/svelte/**/*.test.ts'
+			// 'packages/svelte/tests/*/test.ts',
+			// 'packages/svelte/tests/runtime-browser/test-ssr.ts'
 		],
 		exclude: [...configDefaults.exclude, '**/samples/**'],
 		coverage: {
